[Viz] Guard all HostGMBManager calls to GPUService::DestroyGMB()
In preparation for having HostGpuMemoryBufferManager allocate
SharedMemory GMBs via GPUService rather than doing so itself, this CL
guards calls to GPUService::DestroyGpuMemoryBuffer() in
HostGpuMemoryBufferManager::OnAllocatedGpuMemoryBuffer() with checks
that the GMB is not shared memory. This guard is in place on all other
such calls in HostGMBManager, but to date was not needed here because
this function was only called in response to GPUService having allocated
a native GMB.

Bug: 338958218
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5538371
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1301793}

+++ components/viz/host/host_gpu_memory_buffer_manager.cc
@@ -441,7 +441,7 @@ void HostGpuMemoryBufferManager::OnGpuMemoryBufferAllocated(
   if (client_iter == pending_buffers_.end()) {
     // The client has been destroyed since the allocation request was made. The
     // callback is already called with null handle.
-    if (!handle.is_null()) {
+    if (!handle.is_null() && handle.type != gfx::SHARED_MEMORY_BUFFER) {
       auto* gpu_service = GetGpuService();
       gpu_service->DestroyGpuMemoryBuffer(handle.id, client_id);
     }
@@ -450,7 +450,7 @@ void HostGpuMemoryBufferManager::OnGpuMemoryBufferAllocated(
 
   auto buffer_iter = client_iter->second.find(id);
   if (buffer_iter == client_iter->second.end()) {
-    if (!handle.is_null()) {
+    if (!handle.is_null() && handle.type != gfx::SHARED_MEMORY_BUFFER) {
       // DestroyGpuMemoryBuffer for client_id was called followed by an
       // AllocateGpuMemoryBuffer for a new id.
       auto* gpu_service = GetGpuService();
